modify launch ssh-agent function to reuse existing ssh agent so the pass phrase doesn't have to be entered as often, change vimrc to show whitespace characters (dimly)

+++ .bash_logout
@@ -1,4 +1,5 @@
-# prevents stray ssh-agent processes
-if [ -n "${SSH_AGENT_PID}" ]; then
+# if there are no other bash shells open kill the ssh-agent
+bash_count=$( pgrep -c bash )
+if (( ${bash_count} < 2 )) && [ -n "${SSH_AGENT_PID}" ] ; then
     kill "${SSH_AGENT_PID}"
 fi
